+++ source/terrain/FileUnpacker.h
@@ -40,7 +40,7 @@ public:
 	// UnpackRaw: unpack given number of bytes from the input stream into the given array
 	//	- throws CFileEOFError if the end of the data stream is reached before the given 
 	// number of bytes have been read
-	void UnpackRaw(void* rawdata,u32 rawdatalen);
+	void UnpackRaw(void* rawdata,size_t rawdatalen);
 	// UnpackString: unpack a string from the raw data stream
 	void UnpackString(CStr& result);
